[FE/feat] 홈state_card stateful 클래스로 생성 (#169)
